final template web

+++ app/logout/page.tsx
@@ -13,14 +13,18 @@ const Page = () => {
       if (Cookies.get('isAuthenticated')) {
         await fetchWithCredentials('http://localhost:3001/api/logout');
       }
-      window.localStorage.clear();
+      if (typeof window !== 'undefined') {
+        window.localStorage.clear();
+      }
       Object.keys(Cookies.get()).forEach((cookieName) =>
         Cookies.remove(cookieName),
       );
       setsuccess(true);
     } catch (error) {
       console.error('Error during logout:', error);
-      window.localStorage.clear();
+      if (typeof window !== 'undefined') {
+        window.localStorage.clear();
+      }
       Object.keys(Cookies.get()).forEach((cookieName) =>
         Cookies.remove(cookieName),
       );
@@ -30,12 +34,15 @@ const Page = () => {
 
   if (!success) {
     return (
-      <div className='flex h-screen w-screen items-center justify-center text-xl text-black'>
-        Logging out
+      <div className='flex h-screen w-screen animate-pulse items-center justify-center text-xl font-medium text-black'>
+        Logging you out...
       </div>
     );
   }
-  window.location.href = '/login';
+  if (typeof window !== 'undefined') {
+    window.location.href = '/login';
+  }
+  return null;
 };
 
 export default Page;
